REF rename submodule to synthetic_data (#32)
